proxyFactoryBean 리펙토링
